requirements.txt
beautifulsoup4
feedparser
python-dateutil
requests
sumy
